Apple Picker pt. 5
Created text elements and added code that adds points for every apple that is dropped as well as code to notify apple picker that an apple was dropped. Also added code to delete a basket every time a basket is dropped and stop the game once all baskets are destroyed.